+++ src/main/java/insane96mcp/runeenchanting/data/provider/RERuneTagProvider.java
@@ -58,7 +58,8 @@ public class RERuneTagProvider extends TagsProvider<Rune> {
                 RERunes.CURSE_OF_CORROSION,
                 RERunes.CURSE_OF_UNDEAD,
                 RERunes.CURSE_OF_INACCURACY,
-                RERunes.CURSE_OF_SLIPPERINESS);
+                RERunes.CURSE_OF_SLIPPERINESS,
+                RERunes.CURSE_OF_EXPOSURE);
 
         add(VANISHABLE,
                 RERunes.CURSE_OF_VANISHING,
@@ -68,8 +69,10 @@ public class RERuneTagProvider extends TagsProvider<Rune> {
                 RERunes.RESPIRATION,
                 RERunes.AQUA_AFFINITY,
                 RERunes.ENLIGHTENED,
+                RERunes.SMOKESCREEN,
                 RERunes.CURSE_OF_OBSCURITY,
-                RERunes.CURSE_OF_STATIC_CHARGE);
+                RERunes.CURSE_OF_STATIC_CHARGE,
+                RERunes.CURSE_OF_EXPOSURE);
 
         add(ARMOR_CHEST_ONLY,
                 RERunes.AIR_AFFINITY,